重构,现在目录更清晰了. </subject>
Branch: dev

<type>:
- [ ] Bug fix
- [ ] Bug fix (Test)
- [ ] New feature
- [x] Breaking change
- [ ] Documentation update
- [x] This change requires a documentation update

<body>
重构,现在以lab为目录,更清晰了.
CMake现已改版,使用正则表达式获取子文件夹.
测试宏现由CMake传至test-target.
改用namespace对main进行了覆盖,不再需要main2与测试宏的组合.
<footer>
