+++ .github/pull_request_template.md
@@ -1,13 +1,3 @@
-<!--
- * @Github: https://github.com/Certseeds/CS203_DSAA_template
- * @Organization: SUSTech
- * @Author: nanoseeds
- * @Date: 2020-07-26 16:30:39
- * @LastEditors: nanoseeds
- * @LastEditTime: 2020-07-26 16:32:12
- * @License: CC-BY-NC-SA_V4_0 or any later version 
- -->
-
 # Description
 
 Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.
@@ -16,26 +6,33 @@ Fixes # (issue)
 
 ## Type of change
 
-Please delete options that are not relevant.
+This commit Works for 
+- [ ] Algorithm
+- [ ] The template itself
+
+PS:这个是单选
+
+What kind of this pr?
 
 - [ ] Bug fix (non-breaking change which fixes an issue)
 - [ ] New feature (non-breaking change which adds functionality)
 - [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
-- [ ] This change requires a documentation update
+- [ ] Documentation update
+- [ ] Documentation update-TODO
+- [ ] More testcase (for show fix or just add)
+
+PS:这个是多选
+  
+Please delete options that are not relevant.
+
+PS: 没有勾选的请删掉
 
 # How Has This Been Tested?
 
-Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration
+- [ ] This can pass test in my platform
+- [ ] This can pass test in GitHub Actions
 
-- [ ] Test A
-- [ ] Test B
-
-**Test Configuration**:
-
-+ Firmware version:
-+ Hardware:
-+ Toolchain:
-+ SDK:
+PS: please ensure both of them are ok
 
 # Checklist
 
@@ -45,5 +42,4 @@ Please describe the tests that you ran to verify your changes. Provide instructi
 - [ ] I have made corresponding changes to the documentation
 - [ ] My changes generate no new warnings
 - [ ] I have added tests that prove my fix is effective or that my feature works
-- [ ] New and existing unit tests pass locally with my changes
 - [ ] Any dependent changes have been merged and published in downstream modules
