+++ .github/workflows/pull_request.yml
@@ -10,11 +10,7 @@ name: test for pull_request
 
 # Controls when the action will run. Triggers the workflow on push or pull request
 # events but only for the master branch
-on:
-  pull_request:
-    branches: [ master ]
-  push:
-    branches: [ dev ]
+on: [pull_request,push]
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
   test:
