+++ README.md
@@ -78,23 +78,23 @@
 ## 实际场景
 
 1. A+B: lab_00_A ,测试样例
-    + 这个问题较为简单,见[A+B](./source/lab_00/lab_00_A.cpp).解决起来不复杂.
+    + 这个问题较为简单,见[A+B](lab_00/lab_00_A.cpp).解决起来不复杂.
     + 虽然手工一个一个输入,然后肉眼观察输出.但是如果我们希望严谨的测试,要100组测试数据,难道每次出新版本都要手动输入100次吗?
       显然,有更好的解决方式:使用测试框架.
     + 在本repo,使用`Catch2`测试框架.
       比如,我们有四组数据,第一组,第二组测试边界值,第三组使用随机数测试对偶性与正确性,第四组测试几个手动的随机值.
-      参见[test_for_lab00_A](./test/lab_00/lab_00_A_test.cpp).
+      参见[test_for_lab00_A](lab_00/lab_00_A_test.cpp).
     + 这样一来,我们只需要每次修改完主文件之后,run `CS203_DSAA_template_test`, 对其进行调用,就能验证其在所有的测试用例上的正确性.
       测试的结果也会出现在输出中.
 2. 多个输出值的检查:`Catch::Matchers`
     + 上面的例子里,输出值只是一个值,所以手动检查的难度不大,但是如果目标输出是一个数组,那么手动检查的难度就非常大了.
     + 举例:[Crzay Plan](https://acm.sustech.edu.cn/onlinejudge/problem.php?id=1250),输入可能有1.1*10^6个.  
       这种情况下对这么多值进行直接的观察就很难,所以我们预先将期望的值直接写在测试文件里,用Catch2内置的Matcher比较.
-      见[test_for_lab00_B](./test/lab_00/lab_00_B_test.cpp)的`CHECK_THAT()`部分.
+      见[test_for_lab00_B](lab_00/lab_00_B_test.cpp)的`CHECK_THAT()`部分.
     + 当然,这种情况也只适用于规模比较小的情况,规模再大的话,直接由人手动写在测试文件里也太占空间了.
 3. 输入输出重定向-1:
     + 常见于tree,graph类的问题,debug需要的数据集都比较大,不方便直接写在代码中.
-    + 比如[判断二分图](./source/lab_00/lab_00_C.cpp),一张图可以有几十上百个node,写在内部占用空间太大.
+    + 比如[判断二分图](lab_00/lab_00_C.cpp),一张图可以有几十上百个node,写在内部占用空间太大.
     + 而在这里,使用`CS203_redirect`对象,便可以省去手动输入的方式.
 
     ``` cpp
@@ -118,7 +118,7 @@
     在样例中,编译出的文件在`CS203_DSAA_template/cmake-build-debug`下,所以需要加`./../`
 
 4. 输入输出重定向-2:
-    + 一般来说,题目的输出不会太复杂,但是反例也不是没有.:比如专门考输出的[立体图](./source/lab_00/lab_00_D.cpp)
+    + 一般来说,题目的输出不会太复杂,但是反例也不是没有.:比如专门考输出的[立体图](lab_00/lab_00_D.cpp)
     + 这种情况下,使用c++的重定向输出就可以较为方便的对输入进行处理,同时保存输出方便调试.
 
     ``` cpp
@@ -135,7 +135,7 @@
 
     这样就将标准输出重定向到了01.test.out中,并与01.data.out比对.
     + 至于比较文件之间的差异,可以使用内置的`compareFiles(string path1,string path2)`函数进行比较.
-    参考[文本比对_test_case_2](./test/lab_00/lab_00_D_test.cpp)
+    参考[文本比对_test_case_2](lab_00/lab_00_D_test.cpp)
 
 5. 为什么要将 `读取` `数据处理` `输出` 分开?
     + 便于理清思路,读完题目之后,不管别的,先把数据读入,输出的函数写好,方便后续写作.
