+++ cmake/cross_platform.cmake
@@ -2,6 +2,7 @@ IF (WIN32)
     add_compile_options("$<$<C_COMPILER_ID:MSVC>:/utf-8>")
     add_compile_options("$<$<CXX_COMPILER_ID:MSVC>:/utf-8>")
     set(CMAKE_CXX_FLAGS "/DWIN32 /D_WINDOWS /Wall /GR /EHsc" CACHE INTERNAL "")
+    MESSAGE(WARNING "WIN32's MSVC will make the output change-line symbol to \\r\\n, which make files comp fail")
 ELSEIF (APPLE)
     MESSAGE(STATUS "Now is Apple systens.")
 ELSEIF (UNIX)
