+++ lab_00/lab_00_C.cpp
@@ -65,6 +65,9 @@ then there will be X unrepeated numbers ∈ [0,N-1]
 // id: 785
 // Graph
 // 判断二分图*/
+#ifdef CS203_DSAA_TEST_MACRO
+namespace lab_00_C{
+#endif
 using std::cin;
 using std::cout;
 using std::queue;
@@ -85,20 +88,11 @@ inline int main2();
 
 inline input_type read();
 
-output_type isBipartite(const input_type &data);
+output_type isBipartite(const input_type &graph);
 
 void output(const output_type &data);
 
-#ifndef CS203_DSAA_TEST_MACRO
-#define CS203_DSAA_TEST_MACRO
-
 int main() {
-    return main2();
-}
-
-#endif  // !CS203_DSAA_TEST_MACRO
-
-inline int main2() {
     auto input_data = read();
     auto output_data = isBipartite(input_data);
     output(output_data);
@@ -166,3 +160,6 @@ static const auto faster_streams = [] {
     // 关闭c++风格输入输出 , 与C风格输入输出的同步,提高性能.
     return 0;
 }();
+#ifdef CS203_DSAA_TEST_MACRO
+}
+#endif
