--- script/CMakeLists_main_template.txt
@@ -1,26 +1,0 @@
-cmake_minimum_required(VERSION 3.10.2)
-project(CS203_DSAA_template)
-
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_FLAGS "${{CMAKE_CXX_FLAGS}} -Wall -Wextra")
-set(CMAKE_CXX_FLAGS_DEBUG "${{CMAKE_CXX_FLAGS_DEBUG}} -O0")
-set(CMAKE_CXX_FLAGS_RELEASE "${{CMAKE_CXX_FLAGS_RELEASE}} -O3")
-set(CMAKE_VERBOSE_MAKEFILE OFF)
-
-string(TIMESTAMP time_str "%Y/%m/%d \n %y:%m:%S")
-MESSAGE("Compiler In ${{time_str}}")
-
-include_directories(${{CMAKE_SOURCE_DIR}}/includes)
-include_directories(${{CMAKE_SOURCE_DIR}}/third_party)
-
-add_subdirectory(${{CMAKE_SOURCE_DIR}}/algorithm)
-
-enable_testing()
-set(dependencies {0})
-foreach (elementName IN LISTS dependencies)
-    add_subdirectory(${{CMAKE_SOURCE_DIR}}/source/lab_${{elementName}})
-    add_subdirectory(${{CMAKE_SOURCE_DIR}}/test/lab_${{elementName}})
-endforeach ()
-unset(dependencies)
-
-
