+++ script/cpp_template.txt
@@ -13,6 +13,10 @@
 #include <unordered_map>
 #include <unordered_set>
 
+#ifdef CS203_DSAA_TEST_MACRO
+namespace lab_{0}_{1}{{
+#endif
+
 using std::cin;
 using std::tie;
 using std::cout;
@@ -28,13 +32,11 @@ using std::vector;
 using std::unordered_map;
 using std::unordered_set;
 using std::priority_queue;
-static constexpr const char end{'\n'};
+static constexpr const char end{{'\n'}};
 
 using num_t = int32_t;
 using input_type = tuple<num_t, num_t>;
 using output_type = num_t;
-
-inline int main2();
 
 inline input_type read();
 
@@ -42,38 +44,30 @@ output_type cal(input_type data);
 
 void output(const output_type &data);
 
-#ifndef CS203_DSAA_TEST_MACRO
-#define CS203_DSAA_TEST_MACRO
-
-int main() {
-    return main2();
-}
-#endif // !CS203_DSAA_TEST_MACRO
-
-inline int main2() {
+int main() {{
     auto input_data = read();
     auto output_data = cal(input_data);
     output(output_data);
     return 0;
-}
+}}
 
-inline input_type read() {
-    num_t a{0},b{0};
+inline input_type read() {{
+    num_t a{{0}},b{{0}};
     std::cin >> a >> b;
     return std::make_tuple(a, b);
-}
+}}
 
-output_type cal(input_type data) {
-    num_t a{0},b{0};
+output_type cal(input_type data) {{
+    num_t a{{0}},b{{0}};
     tie(a, b) = data;
     num_t c = a + b;
     return c;
-}
+}}
 
-void output(const output_type &data) {
+void output(const output_type &data) {{
     cout << data << end;
-}
-static const auto faster_streams = [] {
+}}
+static const auto faster_streams = [] {{
     srand(time(nullptr));
     // use time to init the random seed
     std::ios::sync_with_stdio(false);
@@ -83,4 +77,7 @@ static const auto faster_streams = [] {
     std::cout.tie(nullptr);
     // 关闭c++风格输入输出 , 与C风格输入输出的同步,提高性能.
     return 0;
-}();
+}}();
+#ifdef CS203_DSAA_TEST_MACRO
+}}
+#endif
