+++ script/cpp_test_template.txt
@@ -1,11 +1,15 @@
-#ifndef CS203_DSAA_TEST_MACRO
-#define CS203_DSAA_TEST_MACRO
+#ifdef CS203_DSAA_TEST_MACRO
 
 #include <tuple>
 #include <vector>
 #include <iostream>
-#include "catch_main.hpp"
+#include <catch_main.hpp>
+
 #include "lab_{0}_{1}.cpp"
+
+string CS203_redirect::file_paths = "./../../lab_{0}/lab_{0}_{1}_data/";
+
+namespace lab_{0}_{1}{{
 
 using std::tie;
 using std::cin;
@@ -16,7 +20,6 @@ using std::vector;
 using Catch::Matchers::Equals;
 using Catch::Matchers::UnorderedEquals;
 using Catch::Matchers::Contains;
-string CS203_redirect::file_paths = "./../../../test/lab_{0}/lab_{0}_{1}_data/";
 
 TEST_CASE("test case 1", "[test {0} {1}]") {{
     const auto output_data = cal(std::make_tuple(114, 514));
@@ -31,7 +34,7 @@ TEST_CASE("test case 1", "[test {0} {1}]") {{
 }}
 // 因为[.],所以下面这个被隐藏了,确保需要重定向输入输出时,请删除`[.]`
 TEST_CASE("test case with sequence", "[test {0} {1}][.]") {{
-    const CS203_sequence sequence{{1, 1, 0}}; // // 基础设定,[1,1]
+    CS203_sequence sequence{{1, 1, 0}}; // // 基础设定,[1,1]
     sequence.set_postfix_of_datain("data.in"); // 输入数据后缀,默认为 data.in
     sequence.set_postfix_of_dataout("data.out"); // except输出数据后缀,默认为 data.out
     sequence.set_postfix_of_testout("test.out"); // 测试输出数据后缀,默认为 test.out
@@ -43,10 +46,11 @@ TEST_CASE("test case with sequence", "[test {0} {1}][.]") {{
         tie(datain, dataout, testout) = file_name; // 解包
         {{
             CS203_redirect cr{{datain, testout}}; // 重定向输入,输出
-            main2();
+            main();
             // 用括号括住是为了让CS203_redirect在这里被析构,停止重定向
         }}
         CHECK(compareFiles(testout, dataout));
     }}
+}}
 }}
 #endif //CS203_DSAA_TEST_MACRO
