--- test/lab_00/CMakeLists.txt
@@ -1,16 +1,0 @@
-cmake_minimum_required(VERSION 3.10.2)
-set(PROJECT_ORDER lab_00)
-project(CS203_${PROJECT_ORDER}_test)
-
-set(CMAKE_CXX_STANDARD 11)
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
-include_directories(${CMAKE_SOURCE_DIR}/source/${PROJECT_ORDER})
-set(dependencies A B C D)
-
-enable_testing()
-foreach (elementName IN LISTS dependencies)
-    add_executable(${PROJECT_NAME}_${elementName} ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_ORDER}_${elementName}_test.cpp)
-    MESSAGE(STATUS "${PROJECT_NAME}_${elementName} from ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_ORDER}_${elementName}_test.cpp")
-    add_test(${PROJECT_NAME}_${elementName}_CTEST ${PROJECT_NAME}_${elementName})
-endforeach ()
-unset(dependencies)
