--- test/lab_00/lab_00_D_data/01.test.out
@@ -1,13 +1,0 @@
-......+---+---+...+---+
-..+---+  /   /|../   /|
-./   /|-+---+ |.+---+ |
-+---+ |/   /| +-|   | +
-|   | +---+ |/+---+ |/|
-|   |/   /| +/   /|-+ |
-+---+---+ |/+---+ |/| +
-|   |   | +-|   | + |/.
-|   |   |/  |   |/| +..
-+---+---+---+---+ |/...
-|   |   |   |   | +....
-|   |   |   |   |/.....
-+---+---+---+---+......
